requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
greenlet==1.1.2
idna==3.3
mega.py==1.0.8
onedrivesdk-fork==1.1.14
pathlib==1.0.1
playwright==1.22.0
pycryptodome==3.14.1
pyee==8.1.0
pyfiglet==0.8.post1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tabulate==0.8.9
tenacity==5.1.5
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
websockets==10.1
